Fixed integration

+++ src/integrationTest/java/uk/gov/hmcts/darts/transcriptions/controller/TranscriptionControllerRequestTranscriptionIntTest.java
@@ -137,12 +137,12 @@ class TranscriptionControllerRequestTranscriptionIntTest extends IntegrationBase
             .andExpect(status().isOk())
             .andReturn();
 
-        Long transcriptionId = JsonPath.parse(mvcResult.getResponse().getContentAsString())
+        Integer transcriptionId = JsonPath.parse(mvcResult.getResponse().getContentAsString())
             .read("$.transcription_id");
         assertNotNull(transcriptionId);
 
         TranscriptionRepository transcriptionRepository = dartsDatabase.getTranscriptionRepository();
-        TranscriptionEntity transcriptionEntity = transcriptionRepository.findById(transcriptionId).orElseThrow();
+        TranscriptionEntity transcriptionEntity = transcriptionRepository.findById(transcriptionId.longValue()).orElseThrow();
         assertTrue(transcriptionEntity.getIsManualTranscription());
         assertThat(transcriptionEntity.getCourtroom().getId()).isEqualTo(hearing.getCourtroom().getId());
         List<TranscriptionWorkflowEntity> transcriptionWorkflowEntities = transcriptionEntity.getTranscriptionWorkflowEntities();
@@ -228,7 +228,7 @@ class TranscriptionControllerRequestTranscriptionIntTest extends IntegrationBase
         mockMvc.perform(requestBuilder)
             .andExpect(status().isConflict())
             .andExpect(jsonPath("$.type", is("TRANSCRIPTION_107")))
-            .andExpect(jsonPath("$.duplicate_transcription_id", is(dupeTranscription.getId())));
+            .andExpect(jsonPath("$.duplicate_transcription_id", is(dupeTranscription.getId().intValue())));
     }
 
     @Test
@@ -254,7 +254,7 @@ class TranscriptionControllerRequestTranscriptionIntTest extends IntegrationBase
         mockMvc.perform(requestBuilder)
             .andExpect(status().isConflict())
             .andExpect(jsonPath("$.type", is("TRANSCRIPTION_107")))
-            .andExpect(jsonPath("$.duplicate_transcription_id", is(dupeTranscription.getId())));
+            .andExpect(jsonPath("$.duplicate_transcription_id", is(dupeTranscription.getId().intValue())));
     }
